agent/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "text-to-sql-agent"
version = "0.1.0"
requires-python = ">=3.10"

dependencies = [
    "bedrock-agentcore >= 1.0.3",
    "strands-agents >= 1.13.0",
    "psycopg2-binary >= 2.9.0",
    "boto3 >= 1.35.0",
]
